SONAR-6490 Refactor PermissionsWsCommons
